revert redacted_definition field
