+++ proto/catalog.proto
@@ -100,7 +100,6 @@ message Source {
 
   optional uint64 initialized_at_epoch = 15;
   optional uint64 created_at_epoch = 16;
-  optional string redacted_definition = 17;
 
   // Per-source catalog version, used by schema change.
   uint64 version = 100;
@@ -150,7 +149,6 @@ message Sink {
 
   // Target table id (only applicable for table sink)
   optional uint32 target_table = 21;
-  optional string redacted_definition = 22;
 }
 
 message Connection {
@@ -296,8 +294,6 @@ message Table {
 
   // This field is used to mark the the sink into this table.
   repeated uint32 incoming_sinks = 34;
-
-  optional string redacted_definition = 35;
 
   // Per-table catalog version, used by schema change. `None` for internal tables and tests.
   // Not to be confused with the global catalog version for notification service.
