+++ src/connector/src/sink/catalog/mod.rs
@@ -301,8 +301,6 @@ pub struct SinkCatalog {
     pub sink_from_name: String,
 
     pub target_table: Option<TableId>,
-
-    pub redacted_definition: Option<String>,
 }
 
 impl SinkCatalog {
@@ -341,7 +339,6 @@ impl SinkCatalog {
             sink_from_name: self.sink_from_name.clone(),
             stream_job_status: PbStreamJobStatus::Creating.into(),
             target_table: self.target_table.map(|table_id| table_id.table_id()),
-            redacted_definition: self.redacted_definition.clone(),
         }
     }
 
@@ -431,7 +428,6 @@ impl From<PbSink> for SinkCatalog {
             db_name: pb.db_name,
             sink_from_name: pb.sink_from_name,
             target_table: pb.target_table.map(TableId::new),
-            redacted_definition: pb.redacted_definition,
         }
     }
 }
