+++ src/frontend/src/catalog/source_catalog.rs
@@ -39,7 +39,6 @@ pub struct SourceCatalog {
     pub watermark_descs: Vec<WatermarkDesc>,
     pub associated_table_id: Option<TableId>,
     pub definition: String,
-    pub redacted_definition: Option<String>,
     pub connection_id: Option<ConnectionId>,
     pub created_at_epoch: Option<Epoch>,
     pub initialized_at_epoch: Option<Epoch>,
@@ -66,7 +65,6 @@ impl SourceCatalog {
             info: Some(self.info.clone()),
             watermark_descs: self.watermark_descs.clone(),
             definition: self.definition.clone(),
-            redacted_definition: self.redacted_definition.clone(),
             connection_id: self.connection_id,
             initialized_at_epoch: self.initialized_at_epoch.map(|x| x.0),
             created_at_epoch: self.created_at_epoch.map(|x| x.0),
@@ -125,7 +123,6 @@ impl From<&PbSource> for SourceCatalog {
             watermark_descs,
             associated_table_id: associated_table_id.map(|x| x.into()),
             definition: prost.definition.clone(),
-            redacted_definition: prost.redacted_definition.clone(),
             connection_id,
             created_at_epoch: prost.created_at_epoch.map(Epoch::from),
             initialized_at_epoch: prost.initialized_at_epoch.map(Epoch::from),
