+++ src/frontend/src/handler/mod.rs
@@ -24,7 +24,6 @@ use pgwire::pg_server::BoxedError;
 use pgwire::types::{Format, Row};
 use risingwave_common::bail_not_implemented;
 use risingwave_common::error::{ErrorCode, Result};
-use risingwave_connector::source::UPSTREAM_SOURCE_KEY;
 use risingwave_sqlparser::ast::*;
 
 use self::util::{DataChunkToRowSetAdapter, SourceSchemaCompatExt};
@@ -118,7 +117,6 @@ pub struct HandlerArgs {
     pub sql: Arc<str>,
     pub normalized_sql: String,
     pub with_options: WithOptions,
-    pub redacted_sql: Option<String>,
 }
 
 impl HandlerArgs {
@@ -128,7 +126,6 @@ impl HandlerArgs {
             sql,
             with_options: WithOptions::try_from(stmt)?,
             normalized_sql: Self::normalize_sql(stmt),
-            redacted_sql: try_redact_statement(stmt).as_ref().map(Self::normalize_sql),
         })
     }
 
@@ -699,112 +696,5 @@ pub async fn handle(
             comment,
         } => comment::handle_comment(handler_args, object_type, object_name, comment).await,
         _ => bail_not_implemented!("Unhandled statement: {}", stmt),
-    }
-}
-
-/// Returns a redacted and normalized SQL, or None if fails.
-pub fn try_redact_sql(definition: &str) -> Option<String> {
-    let Ok(ast) = risingwave_sqlparser::parser::Parser::parse_sql(definition) else {
-        tracing::error!("failed to parse relation definition");
-        return None;
-    };
-    use itertools::Itertools;
-    let stmt = ast
-        .into_iter()
-        .exactly_one()
-        .expect("should contains only one statement");
-    try_redact_statement(&stmt)
-        .as_ref()
-        .map(HandlerArgs::normalize_sql)
-}
-
-/// Returns a redacted statement, or None if fails.
-///
-/// It redacts properties in create source/create table/create sink, according to `MarkRedaction::marks`.
-///
-/// TODO #14115: redact table name and column names.
-pub fn try_redact_statement(stmt: &Statement) -> Option<Statement> {
-    use mark_redaction::MarkRedaction;
-    use risingwave_connector::match_source_name_str;
-    use risingwave_connector::source::SourceProperties;
-
-    let mut stmt = stmt.clone();
-    let sql_options = match &mut stmt {
-        Statement::CreateSource { stmt } => &mut stmt.with_properties.0,
-        Statement::CreateTable { with_options, .. } => with_options,
-        Statement::CreateSink { stmt } => &mut stmt.with_properties.0,
-        _ => {
-            return Some(stmt);
-        }
-    };
-    let Ok(mut with_properties) =
-        WithOptions::try_from(sql_options.as_slice()).map(WithOptions::into_inner)
-    else {
-        return None;
-    };
-    let Some(connector) = with_properties.remove(UPSTREAM_SOURCE_KEY) else {
-        tracing::error!("Must specify 'connector' in WITH clause");
-        return None;
-    };
-    fn redact_marked_field(options: &mut Vec<SqlOption>, marks: std::collections::HashSet<String>) {
-        for option in options {
-            if marks.contains(&option.name.real_value()) {
-                option.value = Value::SingleQuotedString("[REDACTED]".into());
-            }
-        }
-    }
-    match_source_name_str!(
-        connector.to_lowercase().as_str(),
-        PropType,
-        {
-            let marked = PropType::marks();
-            redact_marked_field(sql_options, marked);
-        },
-        |_| {}
-    );
-    Some(stmt)
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashSet;
-
-    use itertools::Itertools;
-    use mark_redaction::MarkRedaction as MarkRedactionTrait;
-    use mark_redaction_derive::MarkRedaction;
-
-    #[test]
-    fn test_redaction_options() {
-        #[derive(MarkRedaction)]
-        #[expect(dead_code)]
-        struct Bar<T, F> {
-            #[mark_redaction]
-            i_a: F,
-            #[mark_redaction]
-            i_b: T,
-        }
-
-        #[derive(MarkRedaction)]
-        #[expect(dead_code)]
-        struct Foo {
-            #[mark_redaction]
-            a: i32,
-            b: i32,
-            #[mark_redaction(rename = "c_rename", alias = "c_1", alias = "c_2")]
-            c: i32,
-            #[mark_redaction(alias = "d_1")]
-            d: i32,
-            #[mark_redaction(flatten)]
-            #[rustfmt::skip]
-            e: Bar::<i64,i32>,
-            #[mark_redaction(rename = "fields.a", alias = "fields.b")]
-            f: HashSet<String, String>,
-        }
-
-        let marks: HashSet<String> = Foo::marks();
-        let expected_marks = [
-            "a", "c_rename", "c_1", "c_2", "d", "d_1", "i_a", "i_b", "fields.a", "fields.b",
-        ];
-        assert!(marks.iter().sorted().eq(expected_marks.iter().sorted()));
     }
 }
