+++ src/frontend/src/utils/mod.rs
@@ -32,6 +32,8 @@ pub use index_set::*;
 pub(crate) mod group_by;
 pub mod infer_stmt_row_desc;
 pub mod overwrite_options;
+pub mod redaction;
+
 pub use group_by::*;
 pub use overwrite_options::*;
